更新1.1.8版本; 适配tg协议
