+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "nonebot-plugin-splatoon3"
-version = "1.1.7"
+version = "1.1.8"
 description = "一个基于nonebot2框架的splatoon3游戏日程查询插件"
 authors = ["cypas <ayano05@outlook.com>"]
 readme = "README.md"
@@ -17,6 +17,7 @@ playwright="^1.35.0"
 urllib3="1.26.16"
 requests="^2.28.2"
 nonebot-adapter-onebot = "^2.1.3"
+nonebot-adapter-telegram= "^0.1.0b14"
 
 [build-system]
 requires = ["poetry-core"]
